Using a NewRelicQuerySupplier interface w/ DefaultNewRelicQuerySupplier
