+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/monitor/sampling/newrelic/NewRelicMetricSampler.java
@@ -61,6 +61,7 @@ public class NewRelicMetricSampler extends AbstractMetricSampler {
     static final String NEWRELIC_ACCOUNT_ID_CONFIG = "newrelic.account.id";
     static final String NEWRELIC_QUERY_LIMIT_CONFIG = "newrelic.query.limit";
     static final String CLUSTER_NAME_CONFIG = "newrelic.cell.name";
+    static final String NEWRELIC_QUERY_SUPPLIER_CONFIG = "newrelic.query.supplier";
 
     // We make this protected so we can set it during the tests
     protected NewRelicAdapter _newRelicAdapter;
@@ -73,6 +74,8 @@ public class NewRelicMetricSampler extends AbstractMetricSampler {
     // Currently we are hardcoding this in -> later need to make it specific to whatever cluster
     // this cruise control instance is running on
     private static String CLUSTER_NAME = "";
+
+    private NewRelicQuerySupplier _querySupplier;
 
     @Override
     public void configure(Map<String, ?> configs) {
@@ -100,6 +103,12 @@ public class NewRelicMetricSampler extends AbstractMetricSampler {
                     "%s config is required to have the cluster name", CLUSTER_NAME_CONFIG));
         }
         CLUSTER_NAME = (String) configs.get(CLUSTER_NAME_CONFIG);
+
+        _querySupplier = new DefaultNewRelicQuerySupplier();
+
+        // Setup our results to use whichever query supplier as
+        // the supplier for the broker, topic, and partition maps
+        NewRelicQueryResult.setupQuerySupplier(_querySupplier);
     }
 
     /**
@@ -135,7 +144,6 @@ public class NewRelicMetricSampler extends AbstractMetricSampler {
 
         _httpClient = HttpClients.createDefault();
         _newRelicAdapter = new NewRelicAdapter(_httpClient, endpoint, accountId, apiKey);
-
     }
 
     /**
@@ -172,7 +180,7 @@ public class NewRelicMetricSampler extends AbstractMetricSampler {
      */
     private void runBrokerQueries(ResultCounts counts) throws Exception {
         // Run our broker query first
-        final String brokerQuery = NewRelicQuerySupplier.brokerQuery(CLUSTER_NAME);
+        final String brokerQuery = _querySupplier.brokerQuery(CLUSTER_NAME);
         final List<NewRelicQueryResult> brokerResults = _newRelicAdapter.runQuery(brokerQuery);
 
         for (NewRelicQueryResult result : brokerResults) {
@@ -449,10 +457,10 @@ public class NewRelicMetricSampler extends AbstractMetricSampler {
         // When every broker is in one bin, we don't need to include the "WHERE broker IN"
         // and can just select every broker
         if (queryBins.size() == 1) {
-            queries.add(NewRelicQuerySupplier.topicQuery("", CLUSTER_NAME));
+            queries.add(_querySupplier.topicQuery("", CLUSTER_NAME));
         } else {
             for (NewRelicQueryBin queryBin : queryBins) {
-                queries.add(NewRelicQuerySupplier.topicQuery(queryBin.generateStringForQuery(), CLUSTER_NAME));
+                queries.add(_querySupplier.topicQuery(queryBin.generateStringForQuery(), CLUSTER_NAME));
             }
         }
         return queries;
@@ -469,7 +477,7 @@ public class NewRelicMetricSampler extends AbstractMetricSampler {
     private List<String> getPartitionQueries(List<NewRelicQueryBin> queryBins) {
         List<String> queries = new ArrayList<>();
         for (NewRelicQueryBin queryBin: queryBins) {
-            queries.add(NewRelicQuerySupplier.partitionQuery(queryBin.generateStringForQuery(), CLUSTER_NAME));
+            queries.add(_querySupplier.partitionQuery(queryBin.generateStringForQuery(), CLUSTER_NAME));
         }
         return queries;
     }
